requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
requests>=2.28.0
pdfkit>=1.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0
Pillow>=9.0.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
